sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agent-engineers-sdk"
version = "0.1.0"
description = "SDK for building custom agents for the Agent-Engineers platform"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["agent", "ai", "sdk", "agent-engineers"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = []

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pyyaml>=6.0",
]

[project.scripts]
agent-engineers = "sdk.cli:main"

[tool.setuptools.packages.find]
where = [".."]
include = ["sdk*"]

[tool.pytest.ini_options]
testpaths = ["../tests"]
